TransferInputs test
