+++ raiden-dapp/tests/unit/components/transfer/transfer-menus.spec.ts
@@ -114,7 +114,5 @@ describe('TransferMenus.vue', () => {
         name: RouteNames.CHANNELS,
       })
     );
-
-    console.log(wrapper.html());
   });
 });
